+++ hyperspy/tests/model/test_linear_model.py
@@ -90,7 +90,7 @@ class TestMultiFitLinear:
         cm = pytest.warns(UserWarning) if weighted and not self.s._lazy else \
             dummy_context_manager()
         with cm:
-            m.multifit(optimizer='lstsq', iterpath='serpentine')
+            m.multifit(optimizer='lstsq')
         multi = m.as_signal()
 
         np.testing.assert_allclose(single(), multi())
@@ -102,7 +102,7 @@ class TestMultiFitLinear:
         L.centre.free = L.sigma.free = False
         m.append(L)
 
-        m.multifit(iterpath="serpentine")
+        m.multifit()
         nonlinear = L.A.map.copy()
 
         L.A.map['is_set'] = False
@@ -134,7 +134,7 @@ class TestMultiFitLinear:
         cm = pytest.warns(UserWarning) if weighted and not self.s._lazy else \
             dummy_context_manager()
         with cm:
-            m.multifit(optimizer='lstsq', iterpath='serpentine')
+            m.multifit(optimizer='lstsq')
         multi = m.as_signal()
         # compare fits from first pixel
         np.testing.assert_allclose(single(), multi())
@@ -153,7 +153,7 @@ class TestMultiFitLinear:
         cm = pytest.warns(UserWarning) if weighted and not self.s._lazy else \
             dummy_context_manager()
         with cm:
-           m.multifit(optimizer='lstsq', iterpath='serpentine')
+           m.multifit(optimizer='lstsq')
         multi = m.as_signal()
 
         np.testing.assert_allclose(single(), multi())
@@ -765,7 +765,7 @@ def test_power_law():
     m.set_parameters_not_free(only_nonlinear=True)
     m.plot()
     m.fit(optimizer='lstsq')
-    
+
     np.testing.assert_allclose(pl_ref.A.value, pl.A.value)
     np.testing.assert_allclose(pl_ref.r.value, pl.r.value)
     np.testing.assert_allclose(m.as_signal().data, s.data)
@@ -794,7 +794,7 @@ def test_lorentzian():
     m.set_parameters_not_free(only_nonlinear=True)
     m.plot()
     m.fit(optimizer='lstsq')
-    
+
     np.testing.assert_allclose(l_ref.A.value, l.A.value)
     np.testing.assert_allclose(l_ref.centre.value, l.centre.value)
     np.testing.assert_allclose(l_ref.gamma.value, l.gamma.value)
@@ -805,18 +805,18 @@ def test_lorentzian():
 @pytest.mark.parametrize('nav_dim', (0, 1, 2))
 def test_expression_convolved(nav_dim, multiple_free_parameters):
     s_ref = hs.signals.Signal1D(np.ones(100))
-    
+
     # Create signal to convolve
     to_convolve_component = hs.model.components1D.Gaussian(A=100, sigma=5, centre=10)
     to_convolve = hs.signals.Signal1D(to_convolve_component.function(np.arange(100)))
     to_convolve.axes_manager[-1].offset = -to_convolve_component.centre.value
-    
+
     # Create reference signal from model with convolution
     l_ref = hs.model.components1D.Lorentzian(A=100, centre=20, gamma=4)
     m_ref = s_ref.create_model()
     m_ref.append(l_ref)
     m_ref.low_loss = to_convolve
-    s = m_ref.as_signal()    
+    s = m_ref.as_signal()
 
     if nav_dim >= 1:
         s = hs.stack([s]*2)
@@ -867,7 +867,7 @@ def test_expression_multiple_linear_parameter(nav_dim, convolve):
     if convolve:
         m_ref.low_loss = to_convolve
     s = m_ref.as_signal()
-    
+
     if nav_dim >= 1:
         s = hs.stack([s]*2)
         if convolve:
@@ -876,7 +876,7 @@ def test_expression_multiple_linear_parameter(nav_dim, convolve):
         s = hs.stack([s]*3)
         if convolve:
             to_convolve = hs.stack([to_convolve]*3)
-    
+
     m = s.create_model()
     p = hs.model.components1D.Polynomial(order=2, legacy=False)
     m.append(p)
@@ -886,8 +886,8 @@ def test_expression_multiple_linear_parameter(nav_dim, convolve):
         with pytest.warns(UserWarning):
             m.multifit(optimizer='lstsq')
     else:
-        m.multifit(optimizer='lstsq') 
-    
+        m.multifit(optimizer='lstsq')
+
     np.testing.assert_allclose(p_ref.a0.value, p.a0.value)
     np.testing.assert_allclose(p_ref.a1.value, p.a1.value)
     np.testing.assert_allclose(p_ref.a2.value, p.a2.value)
@@ -906,10 +906,10 @@ def test_multiple_linear_parameters_convolution(nav_dim):
     to_convolve_component = hs.model.components1D.Gaussian(A=1000, sigma=50, centre=100)
     to_convolve = hs.signals.Signal1D(to_convolve_component.function(np.arange(1000)))
     to_convolve.axes_manager[-1].offset = -to_convolve_component.centre.value
-    
+
     l_ref1 = hs.model.components1D.Lorentzian(A=100, centre=200, gamma=10)
     l_ref2 = hs.model.components1D.Lorentzian(A=100, centre=600, gamma=20)
-        
+
     m_ref = s_ref.create_model()
     m_ref.extend([l_ref1, l_ref2])
     m_ref.low_loss = to_convolve