+++ hyperspy/tests/signals/test_assign_subclass.py
@@ -115,10 +115,6 @@ class TestConvertBaseSignal:
         assert isinstance(self.s, _lazy_signals.LazySignal)
         assert self.s._lazy
 
-    def test_set_signal_dimension_deprecation_warning(self):
-        with pytest.warns(VisibleDeprecationWarning):
-            self.s.axes_manager.set_signal_dimension(1)
-
     def test_base_to_1d(self):
         self.s.axes_manager._set_signal_dimension(1)
         self.s._assign_subclass()